MISC: added "benchmark"

+++ peuler55/lychral.cpp
@@ -74,7 +74,7 @@ bool peuler::is_palindrome(uint128_t num)
 		auto number = num.convert_to<std::string>();
 		auto check_len = number.size() / 2;
 		auto s_iter = number.begin(); // Iterator from start of string
-		auto e_iter = s_iter + number.size() - 1; // iterator at end of string
+		auto e_iter = s_iter + static_cast<long long>(number.size()) - 1; // iterator at end of string
 		for (uint64_t i{ 0 };i < check_len;++i) {
 			if(*e_iter!=*s_iter) {
 				return false;
